# dashboard/public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dashboard API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "tags": [
                    "Sign Up"
                ],
                "operationId": "d9a17728f82a002c27ff6fddcc5d94e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "nullable": true,
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "violations": {
                                            "nullable": true,
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "propertyPath": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/profile/name": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "operationId": "ffa146f6e94156a4ac41007eeaf5345f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first",
                                    "last"
                                ],
                                "properties": {
                                    "first": {
                                        "type": "string"
                                    },
                                    "last": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "nullable": true,
                                            "properties": {
                                                "code": {
                                                    "type": "integer"
                                                },
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "violations": {
                                            "nullable": true,
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "propertyPath": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "common"
                        ]
                    },
                    {
                        "scalar": "basicAuth"
                    }
                ]
            }
        },
...
        "/work/projects": {
            "get": {
                "tags": [
                    "Work Projects"
                ],
                "operationId": "a07df57323e00f38ca491a95255aa593",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "count": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "pages": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "common"
                        ]
                    },
                    {
                        "scalar": "basicAuth"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "/authorize",
                        "scopes": {
                            "common": "Common"
                        }
                    }
                }
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}